refactor(drug): group label/AACT helpers under a drug_utils package
Move labels.py and aact_synonyms.py into src/pts/pyspark/drug_utils/ (out of the
pyspark root and common/) so the drug-specific helpers live together. Imports in
chembl_molecule and the tests updated accordingly. No behaviour change.
